Update CompletionDelegate method signature
